Move TexMan.GetCanvas and GetTextureCanvas to common
