+++ src/d_main.cpp
@@ -2926,7 +2926,6 @@ static void Doom_CastSpriteIDToString(FString* a, unsigned int b)
 
 
 extern DThinker* NextToThink;
-extern TArray<FCanvas*> AllCanvases;
 
 static void GC_MarkGameRoots()
 {
@@ -2934,8 +2933,6 @@ static void GC_MarkGameRoots()
 	GC::Mark(staticEventManager.LastEventHandler);
 	for (auto Level : AllLevels())
 		Level->Mark();
-	for (auto canvas : AllCanvases)
-		GC::Mark(canvas);
 
 	// Mark players.
 	for (int i = 0; i < MAXPLAYERS; i++)
