+++ wadsrc/static/zscript/engine/base.zs
@@ -307,6 +307,7 @@ struct TexMan
 	native static int CheckRealHeight(TextureID tex);
 	native static bool OkForLocalization(TextureID patch, String textSubstitute);
 	native static bool UseGamePalette(TextureID tex);
+	native static Canvas GetCanvas(String texture);
 }
 
 /*
